+++ Kokoro/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Improved performance of the Action Costs system.
+
 ## 2.1.0
 Released 17 December 2024.